+++ src/user/use-cases/index.ts
@@ -1,2 +1,2 @@
-export * from './list-by-id/list-by-id';
+export * from './find-by-id/find-by-id';
 export * from './sign-in/sign-in';
